refactor: rename react-node-view-portals to react-node-view-context
Renamed `react-node-view-portals` and related references to `react-node-view-context` for consistency with naming conventions. Updated imports, exports, and context usage across editor components to reflect the changes.
