+++ apps/nextjs/src/components/ui/editor/editors/index.ts
@@ -2,4 +2,4 @@ export * from './editor-container';
 export * from './editor-content';
 export * from './editor-provider';
 export * from './react-node-view';
-export * from './react-node-view-portals';
+export * from './react-node-view-context';
